Add product page mostly compelted
There is a simple validation to check that the user is logged in before they can access the add product page. However any user can access this page. Maybe in the future if we have time we can fix this so that only the seller will have access to this page.